Feat: Create quote form component
This commit creates a quote form component that allows users to
request a free quote for moving services. The form includes fields
for name, email, phone number, move date, move size, current
address, destination address, and additional notes.
