+++ src/app/free-quote/page.tsx
@@ -1,4 +1,4 @@
-import QuoteComponent from "@/components/FreeQuote/QuoteComponent";
+import { QuoteComponent } from "@/components/FreeQuote/QuoteComponent";
 import React from "react";
 
 const FreeQuote = () => {
