+++ src/components/icons/insurance-icon.tsx
@@ -1,33 +1,32 @@
-import type { SVGProps } from "react";
+import type { SVGProps, FC } from "react";
+
+
 interface SVGRProps {
 	title?: string;
 	titleId?: string;
 	desc?: string;
 	descId?: string;
 }
-const SvgComponent = ({
+export const InsuranceIcon: FC<SVGProps<SVGSVGElement> & SVGRProps> = ({
 	title,
 	titleId,
 	desc,
 	descId,
 	...props
-}: SVGProps<SVGSVGElement> & SVGRProps) => (
+}) => (
 	<svg
 		{...props}
 		xmlns="http://www.w3.org/2000/svg"
 		xmlSpace="preserve"
 		width={512}
 		height={512}
-		style={{
-			background: "new 0 0 512 512",
-		}}
 		viewBox="0 0 12.7 12.7"
 		aria-labelledby={titleId}
 		aria-describedby={descId}
 	>
 		{desc ? <desc id={descId}>{desc}</desc> : null}
 		{title ? <title id={titleId}>{title}</title> : null}
-		<g fill="#ffd00d">
+		<g>
 			<path
 				d="M5.294.795a.265.265 0 0 0-.065.01l-3.856 1.08a.794.794 0 0 0-.577.764V3.71a.265.265 0 1 0 .53 0V2.65a.26.26 0 0 1 .19-.254L5.3 1.336l3.783 1.06a.262.262 0 0 1 .192.253v3.712a.265.265 0 1 0 .53 0V2.649a.795.795 0 0 0-.58-.763L5.373.805a.265.265 0 0 0-.078-.01zm0 1.08a.265.265 0 0 0-.067.011l-3.179.91a.265.265 0 0 0-.192.255v4.37c0 .464.224.91.545 1.32.32.41.744.79 1.197 1.114.79.564 1.43.837 1.593.91a.265.265 0 0 0 .217 0c.133-.06.564-.249 1.121-.598a.265.265 0 1 0-.281-.449c-.476.298-.784.432-.954.509a7.621 7.621 0 0 1-1.387-.804c-.42-.3-.812-.653-1.089-1.009-.277-.355-.432-.707-.432-.993V3.25l2.913-.834 2.917.833v3.298a.265.265 0 1 0 .53 0V3.051a.265.265 0 0 0-.193-.255l-3.18-.91a.265.265 0 0 0-.08-.01zm.005 1.57c-.728 0-1.324.596-1.324 1.325 0 .403.186.76.472 1.004a1.86 1.86 0 0 0-1.002 1.647.265.265 0 1 0 .53 0 1.325 1.325 0 1 1 2.651 0 .265.265 0 0 0 .53 0 1.86 1.86 0 0 0-1.004-1.647c.287-.243.474-.6.474-1.004a1.33 1.33 0 0 0-1.327-1.324zm0 .531c.443 0 .797.352.797.794s-.354.794-.797.794-.794-.352-.794-.794.352-.794.794-.794zm-4.243.562a.265.265 0 0 0-.261.268v2.615c0 1.51 1.183 2.58 2.187 3.296 1.005.719 1.88 1.09 2.004 1.143.2.086.427.086.628 0 .094-.04.592-.255 1.24-.644a.265.265 0 0 0-.274-.454 9.971 9.971 0 0 1-1.175.611.266.266 0 0 1-.21 0c-.133-.057-.947-.403-1.905-1.087C2.33 9.6 1.326 8.649 1.326 7.42l-.001-2.615a.265.265 0 0 0-.27-.268z"
 				data-original="#000000"
@@ -41,4 +40,3 @@ const SvgComponent = ({
 		</g>
 	</svg>
 );
-export { SvgComponent as InsuranceIcon };
